Move NumericArray.hpp up a directory.
