+++ src/axom/primal/geometry/NURBSPatch.hpp
@@ -15,7 +15,7 @@
 #include "axom/core.hpp"
 #include "axom/slic.hpp"
 
-#include "axom/core/utilities/NumericArray.hpp"
+#include "axom/core/NumericArray.hpp"
 #include "axom/primal/geometry/Point.hpp"
 #include "axom/primal/geometry/Vector.hpp"
 #include "axom/primal/geometry/Segment.hpp"
