+++ src/axom/primal/geometry/OrientedBoundingBox.hpp
@@ -12,7 +12,7 @@
 #include "axom/core.hpp"
 #include "axom/core/NumericLimits.hpp"
 
-#include "axom/core/utilities/NumericArray.hpp"
+#include "axom/core/NumericArray.hpp"
 #include "axom/primal/geometry/Point.hpp"
 #include "axom/primal/geometry/Vector.hpp"
 #include "axom/primal/geometry/BoundingBox.hpp"
