+++ src/axom/primal/geometry/Polyhedron.hpp
@@ -15,7 +15,7 @@
 #include "axom/core/StackArray.hpp"
 
 #include "axom/primal/geometry/Hexahedron.hpp"
-#include "axom/core/utilities/NumericArray.hpp"
+#include "axom/core/NumericArray.hpp"
 #include "axom/primal/geometry/Octahedron.hpp"
 #include "axom/primal/geometry/Point.hpp"
 #include "axom/primal/geometry/Tetrahedron.hpp"
