+++ src/axom/primal/tests/primal_compute_bounding_box.cpp
@@ -8,7 +8,7 @@
 
 #include "gtest/gtest.h"
 
-#include "axom/core/utilities/NumericArray.hpp"
+#include "axom/core/NumericArray.hpp"
 #include "axom/primal/geometry/Point.hpp"
 #include "axom/primal/geometry/Vector.hpp"
 #include "axom/primal/geometry/OrientedBoundingBox.hpp"
