+++ src/axom/primal/tests/primal_point.cpp
@@ -6,7 +6,7 @@
 #include "gtest/gtest.h"
 
 #include "axom/primal/geometry/Point.hpp"
-#include "axom/core/utilities/NumericArray.hpp"
+#include "axom/core/NumericArray.hpp"
 #include "axom/core/execution/execution_space.hpp"
 #include "axom/core/execution/for_all.hpp"
 #include "axom/slic.hpp"
