+++ src/axom/primal/tests/primal_vector.cpp
@@ -5,7 +5,7 @@
 
 #include "gtest/gtest.h"
 #include "axom/slic.hpp"
-#include "axom/core/utilities/NumericArray.hpp"
+#include "axom/core/NumericArray.hpp"
 
 #include "axom/primal/geometry/Vector.hpp"
 #include "axom/primal/geometry/Point.hpp"
